requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
requests==2.31.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
transformers==4.35.2
torch==2.8.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
pyaudio==0.2.11
numpy==1.26.4
scikit-learn==1.3.2
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
